extend graph_reverse_depends to optionally walk revdeps by depkind
The "Dependencies" class's "graph_reverse_depends" method searched all
dependencies defined for a package with no ability to search only
BDEPEND for example.

This commit adds the ability to search only a specific dependency kind
using the "--depkind" flag. It does not support limiting to a set of
depkinds (BDEPEND and DEPEND but not RDEPEND for example) and does not
display any additional information in the output.